+++ src/constants.rs
@@ -3,6 +3,9 @@
 /// Container image registry and organization prefix.
 pub static MODEL_GRAPH_TOOLS_REPOSITORY: &str = "quay.io/modelgraphtools/model";
 
+/// Wado standalone container image repository used for WildFly analysis.
+pub static WADO_SA_REPOSITORY: &str = "quay.io/wado/wado-sa";
+
 /// A static string that defines the current version of the analyzer.
 pub static ANALYZER_VERSION: &str = "0.1.2";