+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix WildFly analysis using upstream WildFly images instead of wado-sa images
+
 ## [0.1.0] - 2026-04-30
 
 ### Added